Ban Checker
Checks if account is banned in the client login screen.
"Your account has been involved in serious rule breaking."

The plugin can be used to easily build new functionality on.

The script can also be used to check once, or enable loop so it checks until login screen shows the ban text.
To check once use the config Enable.
To loop, use config enabled and loopCheck.
You can check isBanned in the script or isBanned in the config if writeConfig is true.